refactor: 別クラスに切り出し
